requirements.txt
amply==0.1.2
click==7.1.2
config==0.5.0.post0
docutils==0.16
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
psycopg2-binary==2.8.6
PuLP==2.3
pyparsing==2.4.7
Werkzeug==1.0.1
waitress==1.4.4
PyYAML==5.3.1